authentication & internal api calls

+++ app/Repositories/NoteRepository.php
@@ -3,10 +3,14 @@
 namespace App\Repositories;
 
 use App\Exceptions\GeneralException;
+use App\Helpers\UserInternalApi;
+use App\Http\Requests\StoreNoteRequest;
+use App\Http\Requests\UpdateNoteRequest;
 use App\Http\Resources\NoteResource;
 use App\Models\Note;
 use App\Traits\Responses;
 use Illuminate\Http\JsonResponse;
+use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\DB;
 use Symfony\Component\HttpFoundation\Response;
 use Throwable;
@@ -19,13 +23,15 @@ class NoteRepository implements NoteRepositoryInterface {
      * @throws Throwable
      * @throws GeneralException
      */
-    public function create(array $attributes): JsonResponse
+    public function create(StoreNoteRequest $request): JsonResponse
     {
-        return DB::transaction(function () use ($attributes) {
-            $created = Note::query()->create([
-                'title' => data_get($attributes, 'title', 'Untitled'),
-                'content' => data_get($attributes, 'content'),
-                'user_id' => data_get($attributes, 'user_id')
+        $token = $request->bearerToken();
+
+        return DB::transaction(function () use ($request, $token) {
+            $created = Note::create([
+                'title' => $request->input('title'),
+                'content' => $request->input('content'),
+                'user_id' => UserInternalApi::getUserId($token)
             ]);
 
             throw_if(!$created, GeneralException::class, 'Failed to create');
@@ -37,14 +43,13 @@ class NoteRepository implements NoteRepositoryInterface {
      * @throws Throwable
      * @throws GeneralException
      */
-    public function update($model, array $attributes): JsonResponse
+    public function update($model, UpdateNoteRequest $request): JsonResponse
     {
         try {
             \DB::beginTransaction();
             $model->update([
-                'title' => $attributes['title'] ?? $model->title,
-                'content' => $attributes['content'] ?? $model->content,
-                'user_id' => $attributes['user_id'] ?? $model->user_id
+                'title' => $request->input('title') ?? $model->title,
+                'content' => $request->input('content') ?? $model->content,
             ]);
             \DB::commit();
             return $this->success(new NoteResource($model), 'Updated');
@@ -60,14 +65,27 @@ class NoteRepository implements NoteRepositoryInterface {
      */
     public function destroy($model): JsonResponse
     {
+        throw_if(
+            Auth::id() !== $model->user_id,
+            GeneralException::class,
+            "Unauthorized",
+            Response::HTTP_UNAUTHORIZED
+        );
         try {
             \DB::beginTransaction();
             $model->delete();
             \DB::commit();
-            $this->success(null, '', Response::HTTP_NO_CONTENT);
+            return $this->success(null, '', Response::HTTP_NO_CONTENT);
         } catch (\Exception $e) {
             \DB::rollBack();
             throw new GeneralException($e->getMessage(), $e->getCode());
         }
+    }
+
+
+    public function index(): JsonResponse
+    {
+        $notes = Note::where('user_id',  Auth::id())->get();
+        return $this->success(NoteResource::collection($notes), 'Retrieved');
     }
 }
